Added routeValue and getNeighborList functions.
routeValue will take an int array route and an int[][] graph and return
that routes cost. A route looks like {0,1,2,3}. routeValue will add the
final step to finish the loop, without it being included in the route.

getNeighborList will return an ArrayList of all the neighbors for a
given route.

For this problem I swap two adjacent values.
IE:
Original {0,1,2,3}
Neighbors:
{1,0,2,3}
{0,2,1,3}
{0,1,3,2}